tidied up code setup to avoid circular dependency
